feat(#826): delete core/exceptions.py and core/registry.py shims — move all callers to contracts
- Migrate 24 files from `nexus.core.exceptions` to `nexus.contracts.exceptions`
  (including core/metadata.py autogen file)
- Migrate 1 file from `nexus.core.registry` to `nexus.contracts.registry`
- Delete both shim files (91 + 19 lines of pure re-exports)
