+++ src/nexus/core/async_nexus_fs.py
@@ -19,13 +19,13 @@ from pathlib import Path
 from typing import TYPE_CHECKING, Any
 
 from nexus.backends.async_local import AsyncLocalBackend
-from nexus.contracts.types import OperationContext, Permission
-from nexus.core.exceptions import (
+from nexus.contracts.exceptions import (
     ConflictError,
     InvalidPathError,
     NexusFileNotFoundError,
     NexusPermissionError,
 )
+from nexus.contracts.types import OperationContext, Permission
 from nexus.core.metadata import (
     DT_DIR,
     DT_REG,
