--- src/nexus/core/exceptions.py
@@ -1,91 +1,0 @@
-"""Backward-compat re-exports — canonical source is nexus.contracts.exceptions."""
-
-from nexus.contracts.exceptions import (
-    AccessDeniedError,
-    AuditLogError,
-    AuthenticationError,
-    BackendError,
-    BootError,
-    BranchConflictError,
-    BranchError,
-    BranchExistsError,
-    BranchNotFoundError,
-    BranchProtectedError,
-    BranchStateError,
-    CircuitOpenError,
-    ConfigurationError,
-    ConflictError,
-    ConnectorAuthError,
-    ConnectorError,
-    ConnectorQuotaError,
-    ConnectorRateLimitError,
-    DatabaseConnectionError,
-    DatabaseError,
-    DatabaseIntegrityError,
-    DatabaseTimeoutError,
-    InvalidPathError,
-    LockTimeout,
-    MetadataError,
-    NexusError,
-    NexusFileNotFoundError,
-    NexusPermissionError,
-    ParserError,
-    PathNotMountedError,
-    PermissionDeniedError,
-    RemoteConnectionError,
-    RemoteFilesystemError,
-    RemoteTimeoutError,
-    ServiceUnavailableError,
-    StalePointerError,
-    StaleSessionError,
-    UploadChecksumMismatchError,
-    UploadExpiredError,
-    UploadNotFoundError,
-    UploadOffsetMismatchError,
-    ValidationError,
-)
-
-__all__ = [
-    "AccessDeniedError",
-    "AuditLogError",
-    "AuthenticationError",
-    "BackendError",
-    "BootError",
-    "ConfigurationError",
-    "BranchConflictError",
-    "BranchError",
-    "BranchExistsError",
-    "BranchNotFoundError",
-    "BranchProtectedError",
-    "BranchStateError",
-    "CircuitOpenError",
-    "ConflictError",
-    "ConnectorAuthError",
-    "ConnectorError",
-    "ConnectorQuotaError",
-    "ConnectorRateLimitError",
-    "DatabaseConnectionError",
-    "DatabaseError",
-    "DatabaseIntegrityError",
-    "DatabaseTimeoutError",
-    "InvalidPathError",
-    "LockTimeout",
-    "MetadataError",
-    "NexusError",
-    "NexusFileNotFoundError",
-    "NexusPermissionError",
-    "ParserError",
-    "PathNotMountedError",
-    "PermissionDeniedError",
-    "RemoteConnectionError",
-    "RemoteFilesystemError",
-    "RemoteTimeoutError",
-    "ServiceUnavailableError",
-    "StalePointerError",
-    "StaleSessionError",
-    "UploadChecksumMismatchError",
-    "UploadExpiredError",
-    "UploadNotFoundError",
-    "UploadOffsetMismatchError",
-    "ValidationError",
-]
