+++ src/nexus/server/api/v1/routers/events.py
@@ -30,7 +30,7 @@ from fastapi import (
 )
 from fastapi import status as http_status
 
-from nexus.core.exceptions import NexusFileNotFoundError, NexusPermissionError
+from nexus.contracts.exceptions import NexusFileNotFoundError, NexusPermissionError
 from nexus.server.api.v1.dependencies import get_nexus_fs
 from nexus.server.dependencies import get_auth_result, get_operation_context, resolve_auth
