+++ src/nexus/services/mcp_service.py
@@ -191,7 +191,7 @@ class MCPService:
         import asyncio
         import json
 
-        from nexus.core.exceptions import ValidationError
+        from nexus.contracts.exceptions import ValidationError
 
         # Get MCP mount manager
         manager = self._get_mcp_mount_manager()
@@ -318,7 +318,7 @@ class MCPService:
             - Transport is auto-detected: stdio for command, sse for url
             - Tools are automatically synced after mounting
         """
-        from nexus.core.exceptions import ValidationError
+        from nexus.contracts.exceptions import ValidationError
         from nexus.mcp.models import MCPMount
 
         # Validate: need either command or url
@@ -405,7 +405,7 @@ class MCPService:
             for m in mounts:
                 service.mcp_unmount(m['name'], context=context)
         """
-        from nexus.core.exceptions import ValidationError
+        from nexus.contracts.exceptions import ValidationError
 
         manager = self._get_mcp_mount_manager()
 
@@ -458,7 +458,7 @@ class MCPService:
         """
         import asyncio
 
-        from nexus.core.exceptions import ValidationError
+        from nexus.contracts.exceptions import ValidationError
 
         manager = self._get_mcp_mount_manager()
