+++ src/nexus/services/search_listing_mixin.py
@@ -19,8 +19,8 @@ import time
 from concurrent.futures import as_completed
 from typing import TYPE_CHECKING, Any
 
+from nexus.contracts.exceptions import PermissionDeniedError
 from nexus.contracts.types import Permission
-from nexus.core.exceptions import PermissionDeniedError
 from nexus.core.rpc_decorator import rpc_expose
 from nexus.raft.zone_manager import ROOT_ZONE_ID
